+++ Bibliotheque/src/framesComponents/SearchPanel.java
@@ -9,7 +9,7 @@ import javax.swing.JScrollPane;
 import javax.swing.JTable;
 import javax.swing.JTextField;
 import tableModels.TMSearchBook;
-import interfaces.DefineComponents;
+import framesInterfaces.DefineComponents;
 import java.awt.Font;
 import javax.swing.JLabel;
 import javax.swing.JSeparator;
