+++ Bibliotheque/src/framesListeners/AProposListener.java
@@ -1,0 +1,22 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package framesListeners;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+/**
+ *
+ * @author sire_marcos
+ */
+public class AProposListener implements ActionListener{
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+        
+    }
+    
+}
