+++ Bibliotheque/src/framesListeners/QuiterListener.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package listeners;
+package framesListeners;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
