--- Bibliotheque/src/listeners/RechercheListener.java
@@ -1,28 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package listeners;
-
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-/**
- *
- * @author Marcos
- */
-public class RechercheListener implements ActionListener{
-    // Fields
-    
-    // Objects
-    
-    // Constructor
-    public RechercheListener() {
-        
-    }
-
-    @Override
-    public void actionPerformed(ActionEvent e) {
-    }
-}
