+++ Bibliotheque/src/tableModels/TMSearchUser.java
@@ -5,7 +5,7 @@
  */
 package tableModels;
 
-import interfaces.DataTables;
+import framesInterfaces.DataTables;
 import java.util.LinkedList;
 import javax.swing.table.DefaultTableModel;
 import tableData.SearchedUser;
